# webserver/doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "",
        "version": "0.0.2",
        "title": "Linto admin API documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "contact@linto.ai"
        }
    },
    "host": "linto.ai/api",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/api/flow/{flowId}": {
            "delete": {
                "tags": [
                    "flow"
                ],
                "summary": "Delete a flow from nodered api by its flowId",
                "operationId": "DeleteFlowFromBLS",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "flowId",
                    "in": "query",
                    "description": "nodered flow id",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }
                }
            }
        },
...
        "/api/flow/postbls/static": {
            "post": {
                "tags": [
                    "flow"
                ],
                "summary": "Publish a static workflow on BLS",
                "operationId": "postStaticFlowOnBLS",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "payload",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "sn": {
                                "description": "Static client serial number",
                                "type": "string"
                            },
                            "workflow_name": {
                                "description": "static workflow name",
                                "type": "string"
                            },
                            "workflowTemplate": {
                                "description": "Workflow template name that will be used",
                                "type": "string"
                            },
                            "sttServiceLanguage": {
                                "description": "Language of the STT service that will be used",
                                "type": "string"
                            },
                            "sttService": {
                                "description": "STT service name that will be used",
                                "type": "string"
                            },
                            "tockApplicationName": {
                                "description": "Tock application name that will be used",
                                "type": "string"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }
                }
            }
        },
        "/api/flow/postbls/application": {
            "post": {
                "tags": [
                    "flow"
                ],
                "summary": "Publish an application workflow on BLS",
                "operationId": "postApplicationFlowOnBLS",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "payload",
                    "in": "body",
                    "description": "static workflow name",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "workflow_name": {
                                "description": "static workflow name",
                                "type": "string"
                            },
                            "workflowTemplate": {
                                "description": "Workflow template name that will be used",
                                "type": "string"
                            },
                            "sttServiceLanguage": {
                                "description": "Language of the STT service that will be used",
                                "type": "string"
                            },
                            "sttService": {
                                "description": "STT service name that will be used",
                                "type": "string"
                            },
                            "tockApplicationName": {
                                "description": "Tock application name that will be used",
                                "type": "string"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }
                }
            }
        },
...
        "/api/stt/lexicalseeding": {
            "post": {
                "tags": [
                    "stt"
                ],
                "summary": "Trigger the process of lexical seeding on a STT service",
                "operationId": "SttLexicalSeeding",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "payload",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "flowId": {
                                "type": "string"
                            },
                            "service_name": {
                                "type": "string"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }
                }
            }
        },
        "/api/stt/generategraph": {
            "post": {
                "tags": [
                    "stt"
                ],
                "summary": "Trigger the process of graph generation on a STT service",
                "operationId": "SttGenerateGraph",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "serviceName",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }
                }
            }
        },
...
        "/api/tock/lexicalseeding": {
            "post": {
                "tags": [
                    "tock"
                ],
                "summary": "Trigger the process of lexical seeding on a Tock application",
                "operationId": "SttLexicalSeeding",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "flowId",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/server_response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server_response": {

            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "enum": ["success", "error"],
                    "required": true
                },
                "msg": {
                    "type": "string",
                    "required": true
                },
                "error": {
                    "type": "object",
                    "required": false
                }
            }
        }
    }
}